# Conflicts:
#	front/src/pages/MyProfile.js

+++ front/src/components/common/ActionSelectorModal.js
@@ -1,7 +1,8 @@
-import React, { useContext } from "react";
+import React, { useContext, useState } from "react";
 import { ListGroup, Modal } from "react-bootstrap";
 import * as Api from "../../Api";
-import { EditFormContext } from "../../App";
+import { EditFormContext, UserStateContext } from "../../App";
+import ToastWrapper from "./ToastWrapper";
 
 /**
  * @param "취소, 수정, 삭제" 등 유저 입력을 받는 모달창
@@ -11,13 +12,20 @@ const ActionSelectorModal = ({
   isActionModalVisible,
   setIsActionModalVisible,
   reviewId,
+  authorId,
   setReviews,
 }) => {
+  const { user: loggedInUser } = useContext(UserStateContext);
   const { setIsEditFormVisible } = useContext(EditFormContext);
+  const [toastMsg, setToastMsg] = useState("");
 
+  // 결과 토스트 팝업 알림: 삭제 성공 혹은 에러
   const deleteReview = async (reviewId) => {
     if (isActionModalVisible) {
       setIsActionModalVisible(false);
+    }
+    if (loggedInUser._id !== authorId) {
+      return setToastMsg("다른사람의 게시물을 삭제할 수 없습니다");
     }
     try {
       const res = await Api.delete(`reviews/${reviewId}`);
@@ -26,48 +34,60 @@ const ActionSelectorModal = ({
       setReviews((current) => {
         return current.filter((review) => review._id !== reviewId);
       });
+      setToastMsg("게시물이 삭제되었습니다");
     } catch (err) {
-      console.error(err);
-      // to do
+      setToastMsg(err);
     }
   };
+
+  const editReview = () => {
+    if (loggedInUser._id !== authorId) {
+      setIsActionModalVisible(false);
+      return setToastMsg("다른사람의 게시물을 수정할 수 없습니다");
+    }
+    setIsActionModalVisible(false);
+    setIsEditFormVisible(true);
+  };
   return (
-    <Modal
-      show={show}
-      onHide={() => setIsActionModalVisible(false)}
-      backdrop="true"
-      keyboard={false}
-      aria-labelledby="contained-modal-title-vcenter"
-      centered
-    >
-      <ListGroup className="text-center">
-        <ListGroup.Item
-          key="edit"
-          action
-          onClick={() => {
-            setIsActionModalVisible(false);
-            setIsEditFormVisible(true);
-          }}
-        >
-          수정
-        </ListGroup.Item>
-        <ListGroup.Item
-          key="del"
-          action
-          style={{ color: "red" }}
-          onClick={() => deleteReview(reviewId)}
-        >
-          삭제
-        </ListGroup.Item>
-        <ListGroup.Item
-          key="cancel"
-          action
-          onClick={() => setIsActionModalVisible(false)}
-        >
-          취소
-        </ListGroup.Item>
-      </ListGroup>
-    </Modal>
+    <>
+      <Modal
+        show={show}
+        onHide={() => setIsActionModalVisible(false)}
+        backdrop="true"
+        keyboard={false}
+        aria-labelledby="contained-modal-title-vcenter"
+        centered
+      >
+        <ListGroup className="text-center">
+          <ListGroup.Item key="edit" action onClick={editReview}>
+            수정
+          </ListGroup.Item>
+          <ListGroup.Item
+            key="del"
+            action
+            style={{ color: "red" }}
+            onClick={() => deleteReview(reviewId)}
+          >
+            삭제
+          </ListGroup.Item>
+          <ListGroup.Item
+            key="cancel"
+            action
+            onClick={() => setIsActionModalVisible(false)}
+          >
+            취소
+          </ListGroup.Item>
+        </ListGroup>
+      </Modal>
+      {toastMsg && (
+        <ToastWrapper
+          onClose={() => setToastMsg("")}
+          text={toastMsg}
+          bg="warning"
+          position="bottom-center"
+        />
+      )}
+    </>
   );
 };
